# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type BrandSupport = (typeof BrandSupport)[keyof typeof BrandSupport];

export const BrandSupport = {
  full: "full",
  chrome_only: "chrome_only",
} as const;

export interface Game {
  id: string;
  slug: string;
  name: string;
  tagline: string;
  thumbnailUrl?: string | null;
  previewBasePath?: string | null;
  brandSupport: BrandSupport;
  defaultPrimaryColor: string;
  defaultSecondaryColor: string;
  defaultAccentColor: string;
  defaultLogoUrl?: string | null;
  defaultHeading: string;
  priceCents: number;
}

export type QuizOptionScores = {
  ES: number;
  NX: number;
  RX: number;
  RZ: number;
  TX: number;
  TZ: number;
};

/**
 * Lexus Energy Quiz only — one answer choice and how many points it adds to each of the 6 fixed outcome models.
 */
export interface QuizOption {
  text: string;
  scores: QuizOptionScores;
}

/**
 * Lexus Energy Quiz only — one question and its answer options.
 */
export interface QuizQuestion {
  text: string;
  options: QuizOption[];
}

export interface UpsertBrandingDraftRequest {
  gameId: string;
  draftToken: string;
  primaryColor: string;
  secondaryColor: string;
  accentColor: string;
  logoDataUrl?: string | null;
  bgDataUrl?: string | null;
  fontDataUrl?: string | null;
  quiz?: QuizQuestion[] | null;
  brandName?: string | null;
  heading: string;
  tagline?: string | null;
}

export type BrandingDraftStatus =
  (typeof BrandingDraftStatus)[keyof typeof BrandingDraftStatus];

export const BrandingDraftStatus = {
  draft: "draft",
  finalized: "finalized",
} as const;

export interface BrandingDraft {
  id: string;
  gameId: string;
  draftToken: string;
  primaryColor: string;
  secondaryColor: string;
  accentColor: string;
  logoDataUrl?: string | null;
  bgDataUrl?: string | null;
  fontDataUrl?: string | null;
  quiz?: QuizQuestion[] | null;
  brandName?: string | null;
  heading: string;
  tagline?: string | null;
  status: BrandingDraftStatus;
}

export interface FinalizeBrandingDraftRequest {
  contactEmail?: string | null;
}

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus];

export const OrderStatus = {
  pending_payment: "pending_payment",
} as const;

export interface Order {
  id: string;
  brandingDraftId: string;
  gameId: string;
  status: OrderStatus;
  totalAmountCents: number;
  contactEmail?: string | null;
}

export interface CreateCustomUiRequestRequest {
  gameId?: string | null;
  name: string;
  email: string;
  message: string;
}

export interface CustomUiRequest {
  id: string;
  gameId?: string | null;
  name: string;
  email: string;
  message: string;
}

export type CreatePlayerSubmissionRequestExtra = {
  [key: string]: unknown;
} | null;

export interface CreatePlayerSubmissionRequest {
  gameSlug: string;
  orderId?: string | null;
  name: string;
  email: string;
  extra?: CreatePlayerSubmissionRequestExtra;
}

export type PlayerSubmissionExtra = { [key: string]: unknown } | null;

export interface PlayerSubmission {
  id: string;
  gameId: string;
  orderId?: string | null;
  name: string;
  email: string;
  extra?: PlayerSubmissionExtra;
}

export interface Scene {
  id: string;
  name: string;
  tagline: string;
  description: string;
}

export interface GenerationInput {
  /** @minLength 1 */
  sceneId: string;
  /**
   * Data URI or raw base64 of the visitor's captured photo
   * @minLength 1
   */
  photoBase64: string;
  visitorName?: string;
}

export interface Generation {
  id: string;
  sceneId: string;
  sceneName: string;
  visitorName?: string | null;
  /** One of pending, ready, failed */
  photoStatus: string;
  /** One of idle, pending, ready, failed */
  videoStatus: string;
  photoUrl?: string | null;
  videoUrl?: string | null;
  error?: string | null;
  createdAt: string;
}

export interface SceneCount {
  sceneId: string;
  sceneName: string;
  count: number;
}

export interface BoothStats {
  totalCreations: number;
  videosCreated: number;
  sceneBreakdown: SceneCount[];
}

export interface ApiError {
  error: string;
}

export type FindBrandingDraftParams = {
  gameId: string;
  draftToken: string;
};

export type ListGenerationsParams = {
  /**
   * @minimum 1
   * @maximum 50
   */
  limit?: number;
};