# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /branding-drafts:
    post:
      operationId: upsertBrandingDraft
      tags: [branding-drafts]
      summary: Create or update a branding draft for a game
      description: Upserts by the combination of gameId + draftToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertBrandingDraftRequest"
      responses:
        "200":
          description: The saved branding draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrandingDraft"
    get:
      operationId: findBrandingDraft
      tags: [branding-drafts]
      summary: Look up an in-progress branding draft for a game + draft token
      description: Lets the customize page restore a customer's unfinished customization after a page reload, since draftToken (stored in localStorage) is the only thing it has on mount.
      parameters:
        - name: gameId
          in: query
          required: true
          schema:
            type: string
        - name: draftToken
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The in-progress draft, if one exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrandingDraft"
        "404":
          description: No in-progress draft for this game + draft token
  /branding-drafts/{id}:
    get:
      operationId: getBrandingDraft
      tags: [branding-drafts]
      summary: Get a branding draft by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The branding draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BrandingDraft"
        "404":
          description: Draft not found
  /branding-drafts/{id}/finalize:
    post:
      operationId: finalizeBrandingDraft
      tags: [branding-drafts]
      summary: Finalize a branding draft and create a pending order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FinalizeBrandingDraftRequest"
      responses:
        "200":
          description: The created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Draft not found
...
components:
  schemas:
...
    UpsertBrandingDraftRequest:
      type: object
      properties:
        gameId:
          type: string
        draftToken:
          type: string
        primaryColor:
          type: string
        secondaryColor:
          type: string
        accentColor:
          type: string
        logoDataUrl:
          type: string
          nullable: true
        bgDataUrl:
          type: string
          nullable: true
        fontDataUrl:
          type: string
          nullable: true
        quiz:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/QuizQuestion"
        brandName:
          type: string
          nullable: true
        heading:
          type: string
        tagline:
          type: string
          nullable: true
      required:
        - gameId
        - draftToken
        - primaryColor
        - secondaryColor
        - accentColor
        - heading
    BrandingDraft:
      type: object
      properties:
        id:
          type: string
        gameId:
          type: string
        draftToken:
          type: string
        primaryColor:
          type: string
        secondaryColor:
          type: string
        accentColor:
          type: string
        logoDataUrl:
          type: string
          nullable: true
        bgDataUrl:
          type: string
          nullable: true
        fontDataUrl:
          type: string
          nullable: true
        quiz:
          type: array
          nullable: true
          items:
            $ref: "#/components/schemas/QuizQuestion"
        brandName:
          type: string
          nullable: true
        heading:
          type: string
        tagline:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, finalized]
      required:
        - id
        - gameId
        - draftToken
        - primaryColor
        - secondaryColor
        - accentColor
        - heading
        - status
    QuizOption:
      type: object
      description: Lexus Energy Quiz only — one answer choice and how many points it adds to each of the 6 fixed outcome models.
      properties:
        text:
          type: string
        scores:
          type: object
          properties:
            ES:
              type: integer
            NX:
              type: integer
            RX:
              type: integer
            RZ:
              type: integer
            TX:
              type: integer
            TZ:
              type: integer
          required: [ES, NX, RX, RZ, TX, TZ]
      required:
        - text
        - scores
    QuizQuestion:
      type: object
      description: Lexus Energy Quiz only — one question and its answer options.
      properties:
        text:
          type: string
        options:
          type: array
          items:
            $ref: "#/components/schemas/QuizOption"
      required:
        - text
        - options
    FinalizeBrandingDraftRequest:
      type: object
      properties:
        contactEmail:
          type: string
          nullable: true
    Order:
      type: object
      properties:
        id:
          type: string
        brandingDraftId:
          type: string
        gameId:
          type: string
        status:
          type: string
          enum: [pending_payment]
        totalAmountCents:
          type: integer
        contactEmail:
          type: string
          nullable: true
      required:
        - id
        - brandingDraftId
        - gameId
        - status
        - totalAmountCents
...