# lib/api-zod/src/generated/types/brandingDraft.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BrandingDraftStatus } from "./brandingDraftStatus";
import type { QuizQuestion } from "./quizQuestion";

export interface BrandingDraft {
  id: string;
  gameId: string;
  draftToken: string;
  primaryColor: string;
  secondaryColor: string;
  accentColor: string;
  logoDataUrl?: string | null;
  bgDataUrl?: string | null;
  fontDataUrl?: string | null;
  quiz?: QuizQuestion[] | null;
  brandName?: string | null;
  heading: string;
  tagline?: string | null;
  status: BrandingDraftStatus;
}