# lib/api-zod/src/generated/types/quizOption.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { QuizOptionScores } from "./quizOptionScores";

/**
 * Lexus Energy Quiz only — one answer choice and how many points it adds to each of the 6 fixed outcome models.
 */
export interface QuizOption {
  text: string;
  scores: QuizOptionScores;
}