# lib/api-zod/src/generated/types/quizQuestion.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { QuizOption } from "./quizOption";

/**
 * Lexus Energy Quiz only — one question and its answer options.
 */
export interface QuizQuestion {
  text: string;
  options: QuizOption[];
}